Increase native compilation warnings and fix the fallout
We were using -Wall but now use -Weverything and opt out of only
a handful of issues such as C++-98 compat, which we don't need.

The most notable impact is that the compiler now catches
narrowing implicit conversions, which was also the primary
motivation for doing this now as there was known debt in our
unsupervised usage of 64-bit types for sizes.

We now assert on bad buffer sizes *and* retain full checks in
release builds to be completely defensive against overflow.

We now use int32_t for scratch buffer sizes. int64_t remains for
file offsets and uint64_t remains for memory mapped files, but
int32_t for scratch eliminates a bunch of managed casts and
therefore makes it easier to reason about overflow prevention.

INT32_MAX is also the most portable limit on these buffer sizes
because the behavior of functions like read is only implementation
defined when the input size is greater than SSIZE_MAX and ssize_t
is int32_t on the lowest common denominator of modern platforms.

Enums corresponding to platform values are now enum X : T instead
of enum class X : T. The PAL_* names are descriptive enough without
scoping and the implicit conversion to raw integer is actually helpful
to avoid a bunch of casts, which are otherwise needed to interop with
the corresponding native API.

+++ src/Common/src/Interop/Unix/System.Native/Interop.SNPrintF.cs
@@ -27,7 +27,7 @@ internal static partial class Interop
         /// On failure, returns a negative value.
         /// </returns>
         [DllImport(Libraries.SystemNative, SetLastError = true)]
-        internal static unsafe extern int SNPrintF(byte* str, ulong size, string format, string arg1);
+        internal static unsafe extern int SNPrintF(byte* str, int size, string format, string arg1);
 
         /// <summary>
         /// Takes a string and applies a formatting to it to transfor
@@ -48,6 +48,6 @@ internal static partial class Interop
         /// On failure, returns a negative value.
         /// </returns>
         [DllImport(Libraries.SystemNative, SetLastError = true)]
-        internal static unsafe extern int SNPrintF(byte* str, ulong size, string format, int arg1);
+        internal static unsafe extern int SNPrintF(byte* str, int size, string format, int arg1);
     }
 }
